+++ lib/valvat/utils.rb
@@ -1,4 +1,4 @@
-module Valvat
+class Valvat
   module Utils
     
     EU_COUNTRIES = %w(AT BE BG CY CZ DE DK EE ES FI FR GB GR HU IE IT LT LU LV MT NL PL PT RO SE SI SK)
@@ -7,13 +7,16 @@ module Valvat
     def self.split(vat)
       COUNTRY_PATTERN =~ vat 
       result = [$1, $2]
-      result[0] = "GR" if result[0] == "EL"
-      return [nil, nil] unless EU_COUNTRIES.include?(result[0])
-      result
+      iso_country = vat_country_to_iso_country(result[0])
+      EU_COUNTRIES.include?(iso_country) ? result : [nil, nil]
     end
     
     def self.normalize(vat)
       vat.upcase.gsub(/\A\s+|\s+\Z/, "")
+    end
+    
+    def self.vat_country_to_iso_country(vat_country)
+      vat_country == "EL" ? "GR" : vat_country
     end
   end
 end