+++ spec/spec_helper.rb
@@ -20,3 +20,13 @@ class ModelBase
     @attributes[key]
   end
 end
+
+def without_any_web_requests!
+  before(:all) do
+    FakeWeb.clean_registry
+    FakeWeb.allow_net_connect = false
+  end
+  after(:all) do
+    FakeWeb.allow_net_connect = true
+  end
+end